+++ src/app/store/i_store.hpp
@@ -3,23 +3,43 @@
 
 #include <cstdint>
 
+#include "connections/observable.hpp"
+
+class Connection;   // Forward declaration
+
 namespace app
 {
+    /**
+     * @brief Tag struct for indicating that the dirty state of a store has
+     * changed, this is used for notifying subscribers when the dirty state of a
+     * store changes, the TagType is a boolean indicating whether the store is
+     * dirty (has unsaved changes) or not
+     *
+     */
+    struct DirtyChanged
+    {
+        /// Type alias for the type of data
+        using TagType = bool;
+    };
 
     /**
      * @brief Interface for a generic store
      *
      */
-    class IStore
+    class IStore : public Observable<DirtyChanged>
     {
+       private:
+        /// Type alias for the base class
+        using Base = Observable<DirtyChanged>;
+
+        /// Type alias for the change callback function for the dirty state of
+        /// the store
+        using DirtyChangedFn = void (*)(void*, const bool& isDirty);
+
        private:
         /// Flag to indicate if the store has potentially dirty data that needs
         /// to be committed
         bool _isPotentiallyDirty = false;
-
-       protected:
-        void _markPotentiallyDirty();
-        void _clearPotentiallyDirty();
 
        public:
         virtual ~IStore() = default;
@@ -43,7 +63,14 @@ namespace app
          */
         [[nodiscard]] virtual bool isDirty() const = 0;
 
-        [[nodiscard]] bool isPotentiallyDirty() const;
+        Connection subscribeToDirty(DirtyChangedFn func, void* user);
+
+       protected:
+        void _markPotentiallyDirty();
+        void _clearPotentiallyDirty();
+
+       private:
+        void _notifySubscribers(const bool& isDirty);
     };
 
     /**