+++ src/settings/params/param_container.hpp
@@ -12,7 +12,7 @@ namespace settings
      * parameters. It defines a pure virtual function to get the parameters as a
      * tuple, which derived classes must implement.
      */
-    class ParamContainer
+    class ParamContainer final
     {
        private:
         /// The key for the parameter container